Tests refactoring. No functional change.
